[Driver] Move get driver shapes to basic so that deployment does not need modelwrapper
